requirements.txt
aws-cdk-lib==2.33.0
constructs>=10.0.0,<11.0.0
aws-cdk.aws-lambda-python-alpha
beautifulsoup4
requests
dnspython
boto3
pytest
